layers: Move barrier CoreChecks to synch file
