+++ layers/core_checks/buffer_validation.cpp
@@ -860,382 +860,6 @@ bool CoreChecks::ValidateBarriersToImages(const Location &outer_loc, const CMD_B
     return skip;
 }
 
-template <typename Barrier, typename TransferBarrier>
-bool CoreChecks::ValidateQFOTransferBarrierUniqueness(const Location &loc, const CMD_BUFFER_STATE *cb_state, const Barrier &barrier,
-                                                      const QFOTransferBarrierSets<TransferBarrier> &barrier_sets) const {
-    bool skip = false;
-    const char *handle_name = TransferBarrier::HandleName();
-    const char *transfer_type = nullptr;
-    if (!IsTransferOp(barrier)) {
-        return skip;
-    }
-    const TransferBarrier *barrier_record = nullptr;
-    if (cb_state->IsReleaseOp(barrier) && !QueueFamilyIsExternal(barrier.dstQueueFamilyIndex)) {
-        const auto found = barrier_sets.release.find(barrier);
-        if (found != barrier_sets.release.cend()) {
-            barrier_record = &(*found);
-            transfer_type = "releasing";
-        }
-    } else if (cb_state->IsAcquireOp(barrier) && !QueueFamilyIsExternal(barrier.srcQueueFamilyIndex)) {
-        const auto found = barrier_sets.acquire.find(barrier);
-        if (found != barrier_sets.acquire.cend()) {
-            barrier_record = &(*found);
-            transfer_type = "acquiring";
-        }
-    }
-    if (barrier_record != nullptr) {
-        skip |=
-            LogWarning(cb_state->commandBuffer(), TransferBarrier::ErrMsgDuplicateQFOInCB(),
-                       "%s %s queue ownership of %s (%s), from srcQueueFamilyIndex %" PRIu32 " to dstQueueFamilyIndex %" PRIu32
-                       " duplicates existing barrier recorded in this command buffer.",
-                       loc.Message().c_str(), transfer_type, handle_name, report_data->FormatHandle(barrier_record->handle).c_str(),
-                       barrier_record->srcQueueFamilyIndex, barrier_record->dstQueueFamilyIndex);
-    }
-    return skip;
-}
-
-VulkanTypedHandle BarrierTypedHandle(const VkImageMemoryBarrier &barrier) {
-    return VulkanTypedHandle(barrier.image, kVulkanObjectTypeImage);
-}
-
-VulkanTypedHandle BarrierTypedHandle(const VkImageMemoryBarrier2KHR &barrier) {
-    return VulkanTypedHandle(barrier.image, kVulkanObjectTypeImage);
-}
-
-std::shared_ptr<const IMAGE_STATE> BarrierHandleState(const ValidationStateTracker &device_state,
-                                                      const VkImageMemoryBarrier &barrier) {
-    return device_state.Get<IMAGE_STATE>(barrier.image);
-}
-
-std::shared_ptr<const IMAGE_STATE> BarrierHandleState(const ValidationStateTracker &device_state,
-                                                      const VkImageMemoryBarrier2KHR &barrier) {
-    return device_state.Get<IMAGE_STATE>(barrier.image);
-}
-
-VulkanTypedHandle BarrierTypedHandle(const VkBufferMemoryBarrier &barrier) {
-    return VulkanTypedHandle(barrier.buffer, kVulkanObjectTypeBuffer);
-}
-
-VulkanTypedHandle BarrierTypedHandle(const VkBufferMemoryBarrier2KHR &barrier) {
-    return VulkanTypedHandle(barrier.buffer, kVulkanObjectTypeBuffer);
-}
-
-const std::shared_ptr<const BUFFER_STATE> BarrierHandleState(const ValidationStateTracker &device_state,
-                                                             const VkBufferMemoryBarrier &barrier) {
-    return device_state.Get<BUFFER_STATE>(barrier.buffer);
-}
-
-std::shared_ptr<const BUFFER_STATE> BarrierHandleState(const ValidationStateTracker &device_state,
-                                                       const VkBufferMemoryBarrier2KHR &barrier) {
-    return device_state.Get<BUFFER_STATE>(barrier.buffer);
-}
-
-template <typename Barrier, typename TransferBarrier>
-void CoreChecks::RecordBarrierValidationInfo(const Location &loc, CMD_BUFFER_STATE *cb_state, const Barrier &barrier,
-                                             QFOTransferBarrierSets<TransferBarrier> &barrier_sets) {
-    if (IsTransferOp(barrier)) {
-        if (cb_state->IsReleaseOp(barrier) && !QueueFamilyIsExternal(barrier.dstQueueFamilyIndex)) {
-            barrier_sets.release.emplace(barrier);
-        } else if (cb_state->IsAcquireOp(barrier) && !QueueFamilyIsExternal(barrier.srcQueueFamilyIndex)) {
-            barrier_sets.acquire.emplace(barrier);
-        }
-    }
-
-    // 7.7.4: If the values of srcQueueFamilyIndex and dstQueueFamilyIndex are equal, no ownership transfer is performed, and the
-    // barrier operates as if they were both set to VK_QUEUE_FAMILY_IGNORED.
-    const uint32_t src_queue_family = barrier.srcQueueFamilyIndex;
-    const uint32_t dst_queue_family = barrier.dstQueueFamilyIndex;
-    const bool is_ownership_transfer = src_queue_family != dst_queue_family;
-
-    if (is_ownership_transfer) {
-        // Only enqueue submit time check if it is needed. If more submit time checks are added, change the criteria
-        // TODO create a better named list, or rename the submit time lists to something that matches the broader usage...
-        auto handle_state = BarrierHandleState(*this, barrier);
-        const bool mode_concurrent = handle_state ? handle_state->createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT : false;
-        if (!mode_concurrent) {
-            const auto typed_handle = BarrierTypedHandle(barrier);
-            core_error::LocationCapture loc_capture(loc);
-            cb_state->queue_submit_functions.emplace_back(
-                [loc_capture, typed_handle, src_queue_family, dst_queue_family](
-                    const ValidationStateTracker &device_data, const QUEUE_STATE &queue_state, const CMD_BUFFER_STATE &cb_state) {
-                    return ValidateConcurrentBarrierAtSubmit(loc_capture.Get(), device_data, queue_state, cb_state, typed_handle,
-                                                             src_queue_family, dst_queue_family);
-                });
-        }
-    }
-}
-
-// Verify image barrier image state and that the image is consistent with FB image
-template <typename ImgBarrier>
-bool CoreChecks::ValidateImageBarrierAttachment(const Location &loc, CMD_BUFFER_STATE const *cb_state,
-                                                const FRAMEBUFFER_STATE *framebuffer, uint32_t active_subpass,
-                                                const safe_VkSubpassDescription2 &sub_desc, const VkRenderPass rp_handle,
-                                                const ImgBarrier &img_barrier, const CMD_BUFFER_STATE *primary_cb_state) const {
-    using sync_vuid_maps::GetImageBarrierVUID;
-    using sync_vuid_maps::ImageError;
-
-    bool skip = false;
-    const auto *fb_state = framebuffer;
-    assert(fb_state);
-    const auto img_bar_image = img_barrier.image;
-    bool image_match = false;
-    bool sub_image_found = false;  // Do we find a corresponding subpass description
-    VkImageLayout sub_image_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-    uint32_t attach_index = 0;
-    // Verify that a framebuffer image matches barrier image
-    const auto attachment_count = fb_state->createInfo.attachmentCount;
-    for (uint32_t attachment = 0; attachment < attachment_count; ++attachment) {
-        auto view_state = primary_cb_state ? primary_cb_state->GetActiveAttachmentImageViewState(attachment)
-                                           : cb_state->GetActiveAttachmentImageViewState(attachment);
-        if (view_state && (img_bar_image == view_state->create_info.image)) {
-            image_match = true;
-            attach_index = attachment;
-            break;
-        }
-    }
-    if (image_match) {  // Make sure subpass is referring to matching attachment
-        if (sub_desc.pDepthStencilAttachment && sub_desc.pDepthStencilAttachment->attachment == attach_index) {
-            sub_image_layout = sub_desc.pDepthStencilAttachment->layout;
-            sub_image_found = true;
-        }
-        if (!sub_image_found && IsExtEnabled(device_extensions.vk_khr_depth_stencil_resolve)) {
-            const auto *resolve = LvlFindInChain<VkSubpassDescriptionDepthStencilResolve>(sub_desc.pNext);
-            if (resolve && resolve->pDepthStencilResolveAttachment &&
-                resolve->pDepthStencilResolveAttachment->attachment == attach_index) {
-                sub_image_layout = resolve->pDepthStencilResolveAttachment->layout;
-                sub_image_found = true;
-            }
-        }
-        if (!sub_image_found) {
-            for (uint32_t j = 0; j < sub_desc.colorAttachmentCount; ++j) {
-                if (sub_desc.pColorAttachments && sub_desc.pColorAttachments[j].attachment == attach_index) {
-                    sub_image_layout = sub_desc.pColorAttachments[j].layout;
-                    sub_image_found = true;
-                    break;
-                }
-                if (!sub_image_found && sub_desc.pResolveAttachments &&
-                    sub_desc.pResolveAttachments[j].attachment == attach_index) {
-                    sub_image_layout = sub_desc.pResolveAttachments[j].layout;
-                    sub_image_found = true;
-                    break;
-                }
-            }
-        }
-        if (!sub_image_found) {
-            auto img_loc = loc.dot(Field::image);
-            const auto &vuid = GetImageBarrierVUID(img_loc, ImageError::kRenderPassMismatch);
-            skip |=
-                LogError(rp_handle, vuid,
-                         "%s Barrier for %s is not referenced by the VkSubpassDescription for active subpass (%d) of current %s.",
-                         img_loc.Message().c_str(), report_data->FormatHandle(img_bar_image).c_str(), active_subpass,
-                         report_data->FormatHandle(rp_handle).c_str());
-        }
-    } else {  // !image_match
-        auto img_loc = loc.dot(Field::image);
-        const auto &vuid = GetImageBarrierVUID(img_loc, ImageError::kRenderPassMismatch);
-        skip |= LogError(fb_state->framebuffer(), vuid, "%s Barrier for %s does not match an image from the current %s.",
-                         img_loc.Message().c_str(), report_data->FormatHandle(img_bar_image).c_str(),
-                         report_data->FormatHandle(fb_state->framebuffer()).c_str());
-    }
-    if (img_barrier.oldLayout != img_barrier.newLayout) {
-        auto layout_loc = loc.dot(Field::oldLayout);
-        const auto &vuid = GetImageBarrierVUID(layout_loc, ImageError::kRenderPassLayoutChange);
-        skip |= LogError(cb_state->commandBuffer(), vuid,
-                         "%s As the Image Barrier for %s is being executed within a render pass instance, oldLayout must "
-                         "equal newLayout yet they are %s and %s.",
-                         layout_loc.Message().c_str(), report_data->FormatHandle(img_barrier.image).c_str(),
-                         string_VkImageLayout(img_barrier.oldLayout), string_VkImageLayout(img_barrier.newLayout));
-    } else {
-        if (sub_image_found && sub_image_layout != img_barrier.oldLayout) {
-            const LogObjectList objlist(rp_handle, img_bar_image);
-            auto layout_loc = loc.dot(Field::oldLayout);
-            const auto &vuid = GetImageBarrierVUID(layout_loc, ImageError::kRenderPassLayoutChange);
-            skip |= LogError(objlist, vuid,
-                             "%s Barrier for %s is referenced by the VkSubpassDescription for active "
-                             "subpass (%d) of current %s as having layout %s, but image barrier has layout %s.",
-                             layout_loc.Message().c_str(), report_data->FormatHandle(img_bar_image).c_str(), active_subpass,
-                             report_data->FormatHandle(rp_handle).c_str(), string_VkImageLayout(sub_image_layout),
-                             string_VkImageLayout(img_barrier.oldLayout));
-        }
-    }
-    return skip;
-}
-// explictly instantiate so these can be used in core_validation.cpp
-template bool CoreChecks::ValidateImageBarrierAttachment(const Location &loc, CMD_BUFFER_STATE const *cb_state,
-                                                         const FRAMEBUFFER_STATE *framebuffer, uint32_t active_subpass,
-                                                         const safe_VkSubpassDescription2 &sub_desc, const VkRenderPass rp_handle,
-                                                         const VkImageMemoryBarrier &img_barrier,
-                                                         const CMD_BUFFER_STATE *primary_cb_state) const;
-template bool CoreChecks::ValidateImageBarrierAttachment(const Location &loc, CMD_BUFFER_STATE const *cb_state,
-                                                         const FRAMEBUFFER_STATE *framebuffer, uint32_t active_subpass,
-                                                         const safe_VkSubpassDescription2 &sub_desc, const VkRenderPass rp_handle,
-                                                         const VkImageMemoryBarrier2KHR &img_barrier,
-                                                         const CMD_BUFFER_STATE *primary_cb_state) const;
-
-template <typename ImgBarrier>
-void CoreChecks::EnqueueSubmitTimeValidateImageBarrierAttachment(const Location &loc, CMD_BUFFER_STATE *cb_state,
-                                                                 const ImgBarrier &barrier) {
-    // Secondary CBs can have null framebuffer so queue up validation in that case 'til FB is known
-    if ((cb_state->activeRenderPass) && (VK_NULL_HANDLE == cb_state->activeFramebuffer) &&
-        (VK_COMMAND_BUFFER_LEVEL_SECONDARY == cb_state->createInfo.level)) {
-        const auto active_subpass = cb_state->activeSubpass;
-        const auto rp_state = cb_state->activeRenderPass;
-        const auto &sub_desc = rp_state->createInfo.pSubpasses[active_subpass];
-        // Secondary CB case w/o FB specified delay validation
-        auto *this_ptr = this;  // Required for older compilers with c++20 compatibility
-        core_error::LocationCapture loc_capture(loc);
-        const auto render_pass = rp_state->renderPass();
-        cb_state->cmd_execute_commands_functions.emplace_back(
-            [this_ptr, loc_capture, active_subpass, sub_desc, render_pass, barrier](
-                const CMD_BUFFER_STATE &secondary_cb, const CMD_BUFFER_STATE *primary_cb, const FRAMEBUFFER_STATE *fb) {
-                return this_ptr->ValidateImageBarrierAttachment(loc_capture.Get(), &secondary_cb, fb, active_subpass, sub_desc,
-                                                                render_pass, barrier, primary_cb);
-            });
-    }
-}
-
-void CoreChecks::RecordBarriers(Func func_name, CMD_BUFFER_STATE *cb_state, uint32_t bufferBarrierCount,
-                                const VkBufferMemoryBarrier *pBufferMemBarriers, uint32_t imageMemBarrierCount,
-                                const VkImageMemoryBarrier *pImageMemBarriers) {
-    for (uint32_t i = 0; i < bufferBarrierCount; i++) {
-        Location loc(func_name, Struct::VkBufferMemoryBarrier, Field::pBufferMemoryBarriers, i);
-        RecordBarrierValidationInfo(loc, cb_state, pBufferMemBarriers[i], cb_state->qfo_transfer_buffer_barriers);
-    }
-    for (uint32_t i = 0; i < imageMemBarrierCount; i++) {
-        Location loc(func_name, Struct::VkImageMemoryBarrier, Field::pImageMemoryBarriers, i);
-        const auto &img_barrier = pImageMemBarriers[i];
-        RecordBarrierValidationInfo(loc, cb_state, img_barrier, cb_state->qfo_transfer_image_barriers);
-        EnqueueSubmitTimeValidateImageBarrierAttachment(loc, cb_state, img_barrier);
-    }
-}
-
-void CoreChecks::RecordBarriers(Func func_name, CMD_BUFFER_STATE *cb_state, const VkDependencyInfoKHR &dep_info) {
-    for (uint32_t i = 0; i < dep_info.bufferMemoryBarrierCount; i++) {
-        Location loc(func_name, Struct::VkBufferMemoryBarrier2, Field::pBufferMemoryBarriers, i);
-        RecordBarrierValidationInfo(loc, cb_state, dep_info.pBufferMemoryBarriers[i], cb_state->qfo_transfer_buffer_barriers);
-    }
-    for (uint32_t i = 0; i < dep_info.imageMemoryBarrierCount; i++) {
-        Location loc(func_name, Struct::VkImageMemoryBarrier2, Field::pImageMemoryBarriers, i);
-        const auto &img_barrier = dep_info.pImageMemoryBarriers[i];
-        RecordBarrierValidationInfo(loc, cb_state, img_barrier, cb_state->qfo_transfer_image_barriers);
-        EnqueueSubmitTimeValidateImageBarrierAttachment(loc, cb_state, img_barrier);
-    }
-}
-
-template <typename TransferBarrier, typename Scoreboard>
-bool CoreChecks::ValidateAndUpdateQFOScoreboard(const debug_report_data *report_data, const CMD_BUFFER_STATE &cb_state,
-                                                const char *operation, const TransferBarrier &barrier,
-                                                Scoreboard *scoreboard) const {
-    // Record to the scoreboard or report that we have a duplication
-    bool skip = false;
-    auto inserted = scoreboard->emplace(barrier, &cb_state);
-    if (!inserted.second && inserted.first->second != &cb_state) {
-        // This is a duplication (but don't report duplicates from the same CB, as we do that at record time
-        const LogObjectList objlist(cb_state.commandBuffer(), barrier.handle, inserted.first->second->commandBuffer());
-        skip = LogWarning(objlist, TransferBarrier::ErrMsgDuplicateQFOInSubmit(),
-                          "%s: %s %s queue ownership of %s (%s), from srcQueueFamilyIndex %" PRIu32
-                          " to dstQueueFamilyIndex %" PRIu32 " duplicates existing barrier submitted in this batch from %s.",
-                          "vkQueueSubmit()", TransferBarrier::BarrierName(), operation, TransferBarrier::HandleName(),
-                          report_data->FormatHandle(barrier.handle).c_str(), barrier.srcQueueFamilyIndex,
-                          barrier.dstQueueFamilyIndex, report_data->FormatHandle(inserted.first->second->commandBuffer()).c_str());
-    }
-    return skip;
-}
-
-template <typename TransferBarrier>
-bool CoreChecks::ValidateQueuedQFOTransferBarriers(
-    const CMD_BUFFER_STATE &cb_state, QFOTransferCBScoreboards<TransferBarrier> *scoreboards,
-    const GlobalQFOTransferBarrierMap<TransferBarrier> &global_release_barriers) const {
-    bool skip = false;
-    const auto &cb_barriers = cb_state.GetQFOBarrierSets(TransferBarrier());
-    const char *barrier_name = TransferBarrier::BarrierName();
-    const char *handle_name = TransferBarrier::HandleName();
-    // No release should have an extant duplicate (WARNING)
-    for (const auto &release : cb_barriers.release) {
-        // Check the global pending release barriers
-        const auto set_it = global_release_barriers.find(release.handle);
-        if (set_it != global_release_barriers.cend()) {
-            const QFOTransferBarrierSet<TransferBarrier> &set_for_handle = set_it->second;
-            const auto found = set_for_handle.find(release);
-            if (found != set_for_handle.cend()) {
-                skip |= LogWarning(cb_state.commandBuffer(), TransferBarrier::ErrMsgDuplicateQFOSubmitted(),
-                                   "%s: %s releasing queue ownership of %s (%s), from srcQueueFamilyIndex %" PRIu32
-                                   " to dstQueueFamilyIndex %" PRIu32
-                                   " duplicates existing barrier queued for execution, without intervening acquire operation.",
-                                   "vkQueueSubmit()", barrier_name, handle_name, report_data->FormatHandle(found->handle).c_str(),
-                                   found->srcQueueFamilyIndex, found->dstQueueFamilyIndex);
-            }
-        }
-        skip |= ValidateAndUpdateQFOScoreboard(report_data, cb_state, "releasing", release, &scoreboards->release);
-    }
-    // Each acquire must have a matching release (ERROR)
-    for (const auto &acquire : cb_barriers.acquire) {
-        const auto set_it = global_release_barriers.find(acquire.handle);
-        bool matching_release_found = false;
-        if (set_it != global_release_barriers.cend()) {
-            const QFOTransferBarrierSet<TransferBarrier> &set_for_handle = set_it->second;
-            matching_release_found = set_for_handle.find(acquire) != set_for_handle.cend();
-        }
-        if (!matching_release_found) {
-            skip |= LogError(cb_state.commandBuffer(), TransferBarrier::ErrMsgMissingQFOReleaseInSubmit(),
-                             "%s: in submitted command buffer %s acquiring ownership of %s (%s), from srcQueueFamilyIndex %" PRIu32
-                             " to dstQueueFamilyIndex %" PRIu32 " has no matching release barrier queued for execution.",
-                             "vkQueueSubmit()", barrier_name, handle_name, report_data->FormatHandle(acquire.handle).c_str(),
-                             acquire.srcQueueFamilyIndex, acquire.dstQueueFamilyIndex);
-        }
-        skip |= ValidateAndUpdateQFOScoreboard(report_data, cb_state, "acquiring", acquire, &scoreboards->acquire);
-    }
-    return skip;
-}
-
-bool CoreChecks::ValidateQueuedQFOTransfers(const CMD_BUFFER_STATE &cb_state,
-                                            QFOTransferCBScoreboards<QFOImageTransferBarrier> *qfo_image_scoreboards,
-                                            QFOTransferCBScoreboards<QFOBufferTransferBarrier> *qfo_buffer_scoreboards) const {
-    bool skip = false;
-    skip |=
-        ValidateQueuedQFOTransferBarriers<QFOImageTransferBarrier>(cb_state, qfo_image_scoreboards, qfo_release_image_barrier_map);
-    skip |= ValidateQueuedQFOTransferBarriers<QFOBufferTransferBarrier>(cb_state, qfo_buffer_scoreboards,
-                                                                        qfo_release_buffer_barrier_map);
-    return skip;
-}
-
-template <typename TransferBarrier>
-void RecordQueuedQFOTransferBarriers(QFOTransferBarrierSets<TransferBarrier> &cb_barriers,
-                                     GlobalQFOTransferBarrierMap<TransferBarrier> &global_release_barriers) {
-    // Add release barriers from this submit to the global map
-    for (const auto &release : cb_barriers.release) {
-        // the global barrier list is mapped by resource handle to allow cleanup on resource destruction
-        // NOTE: vl_concurrent_ordered_map::find() makes a thread safe copy of the result, so we must
-        // copy back after updating.
-        auto iter = global_release_barriers.find(release.handle);
-        iter->second.insert(release);
-        global_release_barriers.insert_or_assign(release.handle, iter->second);
-    }
-
-    // Erase acquired barriers from this submit from the global map -- essentially marking releases as consumed
-    for (const auto &acquire : cb_barriers.acquire) {
-        // NOTE: We're not using [] because we don't want to create entries for missing releases
-        auto set_it = global_release_barriers.find(acquire.handle);
-        if (set_it != global_release_barriers.end()) {
-            QFOTransferBarrierSet<TransferBarrier> &set_for_handle = set_it->second;
-            set_for_handle.erase(acquire);
-            if (set_for_handle.size() == 0) {  // Clean up empty sets
-                global_release_barriers.erase(acquire.handle);
-            } else {
-                // NOTE: vl_concurrent_ordered_map::find() makes a thread safe copy of the result, so we must
-                // copy back after updating.
-                global_release_barriers.insert_or_assign(acquire.handle, set_for_handle);
-            }
-        }
-    }
-}
-
-void CoreChecks::RecordQueuedQFOTransfers(CMD_BUFFER_STATE *cb_state) {
-    RecordQueuedQFOTransferBarriers<QFOImageTransferBarrier>(cb_state->qfo_transfer_image_barriers, qfo_release_image_barrier_map);
-    RecordQueuedQFOTransferBarriers<QFOBufferTransferBarrier>(cb_state->qfo_transfer_buffer_barriers,
-                                                              qfo_release_buffer_barrier_map);
-}
-
 template <typename ImgBarrier>
 void CoreChecks::TransitionImageLayouts(CMD_BUFFER_STATE *cb_state, uint32_t barrier_count, const ImgBarrier *barriers) {
     // For ownership transfers, the barrier is specified twice; as a release
@@ -1256,11 +880,6 @@ void CoreChecks::TransitionImageLayouts(CMD_BUFFER_STATE *cb_state, uint32_t bar
         }
     }
 }
-// explictly instantiate this template so it can be used in core_validation.cpp
-template void CoreChecks::TransitionImageLayouts(CMD_BUFFER_STATE *cb_state, uint32_t barrier_count,
-                                                 const VkImageMemoryBarrier *barrier);
-template void CoreChecks::TransitionImageLayouts(CMD_BUFFER_STATE *cb_state, uint32_t barrier_count,
-                                                 const VkImageMemoryBarrier2KHR *barrier);
 
 VkImageLayout NormalizeSynchronization2Layout(const VkImageAspectFlags aspect_mask, VkImageLayout layout);
 
@@ -3669,475 +3288,6 @@ bool CoreChecks::ValidateImageBarrierSubresourceRange(const Location &loc, const
     return ValidateImageSubresourceRange(image_state->createInfo.mipLevels, image_state->createInfo.arrayLayers, subresourceRange,
                                          loc.StringFunc().c_str(), loc.StringField().c_str(), "arrayLayers", image_state->image(),
                                          sync_vuid_maps::GetSubResourceVUIDs(loc));
-}
-
-namespace barrier_queue_families {
-using sync_vuid_maps::GetBarrierQueueVUID;
-using sync_vuid_maps::kQueueErrorSummary;
-using sync_vuid_maps::QueueError;
-
-class ValidatorState {
-  public:
-    ValidatorState(const ValidationStateTracker *device_data, LogObjectList &&obj, const core_error::Location &location,
-                   const VulkanTypedHandle &barrier_handle, const VkSharingMode sharing_mode)
-        : device_data_(device_data),
-          objects_(std::move(obj)),
-          loc_(location),
-          barrier_handle_(barrier_handle),
-          sharing_mode_(sharing_mode),
-          limit_(static_cast<uint32_t>(device_data->physical_device_state->queue_family_properties.size())),
-          mem_ext_(IsExtEnabled(device_data->device_extensions.vk_khr_external_memory)) {}
-
-    // Log the messages using boilerplate from object state, and Vu specific information from the template arg
-    // One and two family versions, in the single family version, Vu holds the name of the passed parameter
-    bool LogMsg(QueueError vu_index, uint32_t family, const char *param_name) const {
-        const std::string val_code = GetBarrierQueueVUID(loc_, vu_index);
-        const char *annotation = GetFamilyAnnotation(family);
-        return device_data_->LogError(
-            objects_, val_code, "%s Barrier using %s %s created with sharingMode %s, has %s %" PRIu32 "%s. %s",
-            loc_.Message().c_str(), GetTypeString(), device_data_->report_data->FormatHandle(barrier_handle_).c_str(),
-            GetModeString(), param_name, family, annotation, kQueueErrorSummary.at(vu_index).c_str());
-    }
-
-    bool LogMsg(QueueError vu_index, uint32_t src_family, uint32_t dst_family) const {
-        const std::string val_code = GetBarrierQueueVUID(loc_, vu_index);
-        const char *src_annotation = GetFamilyAnnotation(src_family);
-        const char *dst_annotation = GetFamilyAnnotation(dst_family);
-        return device_data_->LogError(objects_, val_code,
-                                      "%s Barrier using %s %s created with sharingMode %s, has srcQueueFamilyIndex %" PRIu32
-                                      "%s and dstQueueFamilyIndex %" PRIu32 "%s. %s",
-                                      loc_.Message().c_str(), GetTypeString(),
-                                      device_data_->report_data->FormatHandle(barrier_handle_).c_str(), GetModeString(), src_family,
-                                      src_annotation, dst_family, dst_annotation, kQueueErrorSummary.at(vu_index).c_str());
-    }
-
-    // This abstract Vu can only be tested at submit time, thus we need a callback from the closure containing the needed
-    // data. Note that the mem_barrier is copied to the closure as the lambda lifespan exceed the guarantees of validity for
-    // application input.
-    static bool ValidateAtQueueSubmit(const QUEUE_STATE *queue_state, const ValidationStateTracker *device_data,
-                                      uint32_t src_family, uint32_t dst_family, const ValidatorState &val) {
-        auto error_code = QueueError::kSubmitQueueMustMatchSrcOrDst;
-        uint32_t queue_family = queue_state->queueFamilyIndex;
-        if ((src_family != queue_family) && (dst_family != queue_family)) {
-            const std::string val_code = GetBarrierQueueVUID(val.loc_, error_code);
-            const char *src_annotation = val.GetFamilyAnnotation(src_family);
-            const char *dst_annotation = val.GetFamilyAnnotation(dst_family);
-            return device_data->LogError(queue_state->Handle(), val_code,
-                                         "%s Barrier submitted to queue with family index %" PRIu32
-                                         ", using %s %s created with sharingMode %s, has "
-                                         "srcQueueFamilyIndex %" PRIu32 "%s and dstQueueFamilyIndex %" PRIu32 "%s. %s",
-                                         val.loc_.Message().c_str(), queue_family, val.GetTypeString(),
-                                         device_data->report_data->FormatHandle(val.barrier_handle_).c_str(), val.GetModeString(),
-                                         src_family, src_annotation, dst_family, dst_annotation,
-                                         kQueueErrorSummary.at(error_code).c_str());
-        }
-        return false;
-    }
-    // Logical helpers for semantic clarity
-    inline bool KhrExternalMem() const { return mem_ext_; }
-    inline bool IsValid(uint32_t queue_family) const { return (queue_family < limit_); }
-    inline bool IsValidOrSpecial(uint32_t queue_family) const {
-        return IsValid(queue_family) || (mem_ext_ && QueueFamilyIsExternal(queue_family)) ||
-               (queue_family == VK_QUEUE_FAMILY_IGNORED);
-    }
-
-    // Helpers for LogMsg
-    const char *GetModeString() const { return string_VkSharingMode(sharing_mode_); }
-
-    // Descriptive text for the various types of queue family index
-    const char *GetFamilyAnnotation(uint32_t family) const {
-        const char *external = " (VK_QUEUE_FAMILY_EXTERNAL)";
-        const char *foreign = " (VK_QUEUE_FAMILY_FOREIGN_EXT)";
-        const char *ignored = " (VK_QUEUE_FAMILY_IGNORED)";
-        const char *valid = " (VALID)";
-        const char *invalid = " (INVALID)";
-        switch (family) {
-            case VK_QUEUE_FAMILY_EXTERNAL:
-                return external;
-            case VK_QUEUE_FAMILY_FOREIGN_EXT:
-                return foreign;
-            case VK_QUEUE_FAMILY_IGNORED:
-                return ignored;
-            default:
-                if (IsValid(family)) {
-                    return valid;
-                }
-                return invalid;
-        };
-    }
-    const char *GetTypeString() const { return object_string[barrier_handle_.type]; }
-    VkSharingMode GetSharingMode() const { return sharing_mode_; }
-
-  protected:
-    const ValidationStateTracker *device_data_;
-    const LogObjectList objects_;
-    const core_error::Location loc_;
-    const VulkanTypedHandle barrier_handle_;
-    const VkSharingMode sharing_mode_;
-    const uint32_t limit_;
-    const bool mem_ext_;
-};
-
-bool Validate(const CoreChecks *device_data, const CMD_BUFFER_STATE *cb_state, const ValidatorState &val,
-              const uint32_t src_queue_family, const uint32_t dst_queue_family) {
-    bool skip = false;
-
-    const bool mode_concurrent = val.GetSharingMode() == VK_SHARING_MODE_CONCURRENT;
-    const bool src_ignored = QueueFamilyIsIgnored(src_queue_family);
-    const bool dst_ignored = QueueFamilyIsIgnored(dst_queue_family);
-    if (val.KhrExternalMem()) {
-        if (mode_concurrent) {
-            const bool sync2 = device_data->enabled_features.core13.synchronization2 != 0;
-            // this requirement is removed by VK_KHR_synchronization2
-            if (!(src_ignored || dst_ignored) && !sync2) {
-                skip |= val.LogMsg(QueueError::kSrcOrDstMustBeIgnore, src_queue_family, dst_queue_family);
-            }
-            if ((src_ignored && !(dst_ignored || QueueFamilyIsExternal(dst_queue_family))) ||
-                (dst_ignored && !(src_ignored || QueueFamilyIsExternal(src_queue_family)))) {
-                skip |= val.LogMsg(QueueError::kSpecialOrIgnoreOnly, src_queue_family, dst_queue_family);
-            }
-        } else {
-            // VK_SHARING_MODE_EXCLUSIVE
-            if (src_queue_family != dst_queue_family) {
-                if (!val.IsValidOrSpecial(dst_queue_family)) {
-                    skip |= val.LogMsg(QueueError::kSrcAndDstValidOrSpecial, dst_queue_family, "dstQueueFamilyIndex");
-                }
-                if (!val.IsValidOrSpecial(src_queue_family)) {
-                    skip |= val.LogMsg(QueueError::kSrcAndDstValidOrSpecial, src_queue_family, "srcQueueFamilyIndex");
-                }
-            }
-        }
-    } else {
-        // No memory extension
-        if (mode_concurrent) {
-            const bool sync2 = device_data->enabled_features.core13.synchronization2 != 0;
-            // this requirement is removed by VK_KHR_synchronization2
-            if ((!src_ignored || !dst_ignored) && !sync2) {
-                skip |= val.LogMsg(QueueError::kSrcAndDestMustBeIgnore, src_queue_family, dst_queue_family);
-            }
-        } else {
-            // VK_SHARING_MODE_EXCLUSIVE
-            if ((src_queue_family != dst_queue_family) && !(val.IsValid(src_queue_family) && val.IsValid(dst_queue_family))) {
-                skip |= val.LogMsg(QueueError::kSrcAndDstBothValid, src_queue_family, dst_queue_family);
-            }
-        }
-    }
-    return skip;
-}
-}  // namespace barrier_queue_families
-
-bool CoreChecks::ValidateConcurrentBarrierAtSubmit(const Location &loc, const ValidationStateTracker &state_data,
-                                                   const QUEUE_STATE &queue_state, const CMD_BUFFER_STATE &cb_state,
-                                                   const VulkanTypedHandle &typed_handle, uint32_t src_queue_family,
-                                                   uint32_t dst_queue_family) {
-    using barrier_queue_families::ValidatorState;
-    ValidatorState val(&state_data, LogObjectList(cb_state.Handle()), loc, typed_handle, VK_SHARING_MODE_CONCURRENT);
-    return ValidatorState::ValidateAtQueueSubmit(&queue_state, &state_data, src_queue_family, dst_queue_family, val);
-}
-
-// Type specific wrapper for image barriers
-template <typename ImgBarrier>
-bool CoreChecks::ValidateBarrierQueueFamilies(const Location &loc, const CMD_BUFFER_STATE *cb_state, const ImgBarrier &barrier,
-                                              const IMAGE_STATE *state_data) const {
-    // State data is required
-    if (!state_data) {
-        return false;
-    }
-
-    // Create the validator state from the image state
-    barrier_queue_families::ValidatorState val(this, LogObjectList(cb_state->commandBuffer()), loc, state_data->Handle(),
-                                               state_data->createInfo.sharingMode);
-    const uint32_t src_queue_family = barrier.srcQueueFamilyIndex;
-    const uint32_t dst_queue_family = barrier.dstQueueFamilyIndex;
-    return barrier_queue_families::Validate(this, cb_state, val, src_queue_family, dst_queue_family);
-}
-
-// Type specific wrapper for buffer barriers
-template <typename BufBarrier>
-bool CoreChecks::ValidateBarrierQueueFamilies(const Location &loc, const CMD_BUFFER_STATE *cb_state, const BufBarrier &barrier,
-                                              const BUFFER_STATE *state_data) const {
-    // State data is required
-    if (!state_data) {
-        return false;
-    }
-
-    // Create the validator state from the buffer state
-    barrier_queue_families::ValidatorState val(this, LogObjectList(cb_state->commandBuffer()), loc, state_data->Handle(),
-                                               state_data->createInfo.sharingMode);
-    const uint32_t src_queue_family = barrier.srcQueueFamilyIndex;
-    const uint32_t dst_queue_family = barrier.dstQueueFamilyIndex;
-    return barrier_queue_families::Validate(this, cb_state, val, src_queue_family, dst_queue_family);
-}
-
-template <typename Barrier>
-bool CoreChecks::ValidateBufferBarrier(const LogObjectList &objects, const Location &loc, const CMD_BUFFER_STATE *cb_state,
-                                       const Barrier &mem_barrier) const {
-    using sync_vuid_maps::BufferError;
-    using sync_vuid_maps::GetBufferBarrierVUID;
-
-    bool skip = false;
-
-    skip |= ValidateQFOTransferBarrierUniqueness(loc, cb_state, mem_barrier, cb_state->qfo_transfer_buffer_barriers);
-
-    // Validate buffer barrier queue family indices
-    auto buffer_state = Get<BUFFER_STATE>(mem_barrier.buffer);
-    if (buffer_state) {
-        auto buf_loc = loc.dot(Field::buffer);
-        const auto &mem_vuid = GetBufferBarrierVUID(buf_loc, BufferError::kNoMemory);
-        skip |= ValidateMemoryIsBoundToBuffer(cb_state->commandBuffer(), *buffer_state, loc.StringFunc().c_str(), mem_vuid.c_str());
-
-        skip |= ValidateBarrierQueueFamilies(buf_loc, cb_state, mem_barrier, buffer_state.get());
-
-        auto buffer_size = buffer_state->createInfo.size;
-        if (mem_barrier.offset >= buffer_size) {
-            auto offset_loc = loc.dot(Field::offset);
-            const auto &vuid = GetBufferBarrierVUID(offset_loc, BufferError::kOffsetTooBig);
-
-            skip |= LogError(objects, vuid, "%s %s has offset 0x%" PRIx64 " which is not less than total size 0x%" PRIx64 ".",
-                             offset_loc.Message().c_str(), report_data->FormatHandle(mem_barrier.buffer).c_str(),
-                             HandleToUint64(mem_barrier.offset), HandleToUint64(buffer_size));
-        } else if (mem_barrier.size != VK_WHOLE_SIZE && (mem_barrier.offset + mem_barrier.size > buffer_size)) {
-            auto size_loc = loc.dot(Field::size);
-            const auto &vuid = GetBufferBarrierVUID(size_loc, BufferError::kSizeOutOfRange);
-            skip |= LogError(objects, vuid,
-                             "%s %s has offset 0x%" PRIx64 " and size 0x%" PRIx64 " whose sum is greater than total size 0x%" PRIx64
-                             ".",
-                             size_loc.Message().c_str(), report_data->FormatHandle(mem_barrier.buffer).c_str(),
-                             HandleToUint64(mem_barrier.offset), HandleToUint64(mem_barrier.size), HandleToUint64(buffer_size));
-        }
-        if (mem_barrier.size == 0) {
-            auto size_loc = loc.dot(Field::size);
-            const auto &vuid = GetBufferBarrierVUID(size_loc, BufferError::kSizeZero);
-            skip |= LogError(objects, vuid, "%s %s has a size of 0.", loc.Message().c_str(),
-                             report_data->FormatHandle(mem_barrier.buffer).c_str());
-        }
-    }
-
-    if (mem_barrier.srcQueueFamilyIndex == VK_QUEUE_FAMILY_EXTERNAL &&
-        mem_barrier.dstQueueFamilyIndex == VK_QUEUE_FAMILY_EXTERNAL) {
-        auto size_loc = loc.dot(Field::srcQueueFamilyIndex);
-        const auto &vuid = GetBufferBarrierVUID(size_loc, BufferError::kQueueFamilyExternal);
-        skip |= LogError(objects, vuid, "Both srcQueueFamilyIndex and dstQueueFamilyIndex are VK_QUEUE_FAMILY_EXTERNAL.");
-    }
-    return skip;
-}
-
-template <typename Barrier>
-bool CoreChecks::ValidateImageBarrier(const LogObjectList &objects, const Location &loc, const CMD_BUFFER_STATE *cb_state,
-                                      const Barrier &mem_barrier) const {
-    bool skip = false;
-
-    skip |= ValidateQFOTransferBarrierUniqueness(loc, cb_state, mem_barrier, cb_state->qfo_transfer_image_barriers);
-
-    bool is_ilt = true;
-    if (enabled_features.core13.synchronization2) {
-        is_ilt = mem_barrier.oldLayout != mem_barrier.newLayout;
-    }
-
-    if (is_ilt) {
-        if (mem_barrier.newLayout == VK_IMAGE_LAYOUT_UNDEFINED || mem_barrier.newLayout == VK_IMAGE_LAYOUT_PREINITIALIZED) {
-            auto layout_loc = loc.dot(Field::newLayout);
-            const auto &vuid = sync_vuid_maps::GetImageBarrierVUID(loc, sync_vuid_maps::ImageError::kBadLayout);
-            skip |=
-                LogError(cb_state->commandBuffer(), vuid, "%s Image Layout cannot be transitioned to UNDEFINED or PREINITIALIZED.",
-                         layout_loc.Message().c_str());
-        }
-    }
-
-    if (mem_barrier.newLayout == VK_IMAGE_LAYOUT_ATTACHMENT_FEEDBACK_LOOP_OPTIMAL_EXT) {
-        if (!enabled_features.attachment_feedback_loop_layout_features.attachmentFeedbackLoopLayout) {
-            auto layout_loc = loc.dot(Field::newLayout);
-            const auto &vuid = sync_vuid_maps::GetImageBarrierVUID(loc, sync_vuid_maps::ImageError::kBadAttFeedbackLoopLayout);
-            skip |= LogError(cb_state->commandBuffer(), vuid,
-                             "%s Image Layout cannot be transitioned to VK_IMAGE_LAYOUT_ATTACHMENT_FEEDBACK_LOOP_OPTIMAL_EXT if "
-                             "the attachmentFeedbackLoopLayout feature is not enabled",
-                             layout_loc.Message().c_str());
-        }
-    }
-
-    auto image_data = Get<IMAGE_STATE>(mem_barrier.image);
-    if (image_data) {
-        auto image_loc = loc.dot(Field::image);
-
-        skip |= ValidateMemoryIsBoundToImage(cb_state->commandBuffer(), *image_data, loc);
-
-        skip |= ValidateBarrierQueueFamilies(image_loc, cb_state, mem_barrier, image_data.get());
-
-        skip |= ValidateImageAspectMask(image_data->image(), image_data->createInfo.format, mem_barrier.subresourceRange.aspectMask,
-                                        image_data->disjoint, loc.StringFunc().c_str());
-
-        skip |=
-            ValidateImageBarrierSubresourceRange(loc.dot(Field::subresourceRange), image_data.get(), mem_barrier.subresourceRange);
-        skip |= ValidateImageAcquired(*image_data, loc.StringFunc().c_str());
-    }
-    return skip;
-}
-
-bool CoreChecks::ValidateBarriers(const Location &outer_loc, const CMD_BUFFER_STATE *cb_state, VkPipelineStageFlags src_stage_mask,
-                                  VkPipelineStageFlags dst_stage_mask, uint32_t memBarrierCount,
-                                  const VkMemoryBarrier *pMemBarriers, uint32_t bufferBarrierCount,
-                                  const VkBufferMemoryBarrier *pBufferMemBarriers, uint32_t imageMemBarrierCount,
-                                  const VkImageMemoryBarrier *pImageMemBarriers) const {
-    bool skip = false;
-    LogObjectList objects(cb_state->commandBuffer());
-
-    for (uint32_t i = 0; i < memBarrierCount; ++i) {
-        const auto &mem_barrier = pMemBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkMemoryBarrier, Field::pMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier, src_stage_mask, dst_stage_mask);
-    }
-    for (uint32_t i = 0; i < imageMemBarrierCount; ++i) {
-        const auto &mem_barrier = pImageMemBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkImageMemoryBarrier, Field::pImageMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier, src_stage_mask, dst_stage_mask);
-        skip |= ValidateImageBarrier(objects, loc, cb_state, mem_barrier);
-    }
-    {
-        Location loc(outer_loc.function, Struct::VkImageMemoryBarrier);
-        skip |= ValidateBarriersToImages(loc, cb_state, imageMemBarrierCount, pImageMemBarriers);
-    }
-    for (uint32_t i = 0; i < bufferBarrierCount; ++i) {
-        const auto &mem_barrier = pBufferMemBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkBufferMemoryBarrier, Field::pMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier, src_stage_mask, dst_stage_mask);
-        skip |= ValidateBufferBarrier(objects, loc, cb_state, mem_barrier);
-    }
-    return skip;
-}
-
-bool CoreChecks::ValidateDependencyInfo(const LogObjectList &objects, const Location &outer_loc, const CMD_BUFFER_STATE *cb_state,
-                                        const VkDependencyInfoKHR *dep_info) const {
-    bool skip = false;
-
-    if (cb_state->activeRenderPass) {
-        skip |= ValidateRenderPassPipelineBarriers(outer_loc, cb_state, dep_info);
-        if (skip) return true;  // Early return to avoid redundant errors from below calls
-    }
-    for (uint32_t i = 0; i < dep_info->memoryBarrierCount; ++i) {
-        const auto &mem_barrier = dep_info->pMemoryBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkMemoryBarrier2, Field::pMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier);
-    }
-    for (uint32_t i = 0; i < dep_info->imageMemoryBarrierCount; ++i) {
-        const auto &mem_barrier = dep_info->pImageMemoryBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkImageMemoryBarrier2, Field::pImageMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier);
-        skip |= ValidateImageBarrier(objects, loc, cb_state, mem_barrier);
-    }
-    {
-        Location loc(outer_loc.function, Struct::VkImageMemoryBarrier2);
-        skip |= ValidateBarriersToImages(loc, cb_state, dep_info->imageMemoryBarrierCount, dep_info->pImageMemoryBarriers);
-    }
-
-    for (uint32_t i = 0; i < dep_info->bufferMemoryBarrierCount; ++i) {
-        const auto &mem_barrier = dep_info->pBufferMemoryBarriers[i];
-        auto loc = outer_loc.dot(Struct::VkBufferMemoryBarrier2, Field::pBufferMemoryBarriers, i);
-        skip |= ValidateMemoryBarrier(objects, loc, cb_state, mem_barrier);
-        skip |= ValidateBufferBarrier(objects, loc, cb_state, mem_barrier);
-    }
-
-    return skip;
-}
-
-// template to check all original barrier structures
-template <typename Barrier>
-bool CoreChecks::ValidateMemoryBarrier(const LogObjectList &objects, const Location &loc, const CMD_BUFFER_STATE *cb_state,
-                                       const Barrier &barrier, VkPipelineStageFlags src_stage_mask,
-                                       VkPipelineStageFlags dst_stage_mask) const {
-    bool skip = false;
-    assert(cb_state);
-    auto queue_flags = cb_state->GetQueueFlags();
-
-    if (!cb_state->IsAcquireOp(barrier)) {
-        skip |= ValidateAccessMask(objects, loc.dot(Field::srcAccessMask), queue_flags, barrier.srcAccessMask, src_stage_mask);
-    }
-    if (!cb_state->IsReleaseOp(barrier)) {
-        skip |= ValidateAccessMask(objects, loc.dot(Field::dstAccessMask), queue_flags, barrier.dstAccessMask, dst_stage_mask);
-    }
-    return skip;
-}
-
-// template to check all synchronization2 barrier structures
-template <typename Barrier>
-bool CoreChecks::ValidateMemoryBarrier(const LogObjectList &objects, const Location &loc, const CMD_BUFFER_STATE *cb_state,
-                                       const Barrier &barrier) const {
-    bool skip = false;
-    assert(cb_state);
-    auto queue_flags = cb_state->GetQueueFlags();
-
-    skip |= ValidatePipelineStage(objects, loc.dot(Field::srcStageMask), queue_flags, barrier.srcStageMask);
-    if (!cb_state->IsAcquireOp(barrier)) {
-        skip |=
-            ValidateAccessMask(objects, loc.dot(Field::srcAccessMask), queue_flags, barrier.srcAccessMask, barrier.srcStageMask);
-    }
-
-    skip |= ValidatePipelineStage(objects, loc.dot(Field::dstStageMask), queue_flags, barrier.dstStageMask);
-    if (!cb_state->IsReleaseOp(barrier)) {
-        skip |=
-            ValidateAccessMask(objects, loc.dot(Field::dstAccessMask), queue_flags, barrier.dstAccessMask, barrier.dstStageMask);
-    }
-    return skip;
-}
-
-// VkSubpassDependency validation happens when vkCreateRenderPass() is called.
-// Dependencies between subpasses can only use pipeline stages compatible with VK_QUEUE_GRAPHICS_BIT,
-// for external subpasses we don't have a yet command buffer so we have to assume all of them are valid.
-static inline VkQueueFlags SubpassToQueueFlags(uint32_t subpass) {
-    return subpass == VK_SUBPASS_EXTERNAL ? sync_utils::kAllQueueTypes : static_cast<VkQueueFlags>(VK_QUEUE_GRAPHICS_BIT);
-}
-
-bool CoreChecks::ValidateSubpassDependency(const LogObjectList &objects, const Location &in_loc,
-                                           const VkSubpassDependency2 &dependency) const {
-    bool skip = false;
-    Location loc = in_loc;
-    VkMemoryBarrier2KHR converted_barrier;
-    const auto *mem_barrier = LvlFindInChain<VkMemoryBarrier2KHR>(dependency.pNext);
-
-    if (mem_barrier && enabled_features.core13.synchronization2) {
-        if (dependency.srcAccessMask != 0) {
-            skip |= LogError(objects, "UNASSIGNED-CoreChecks-VkSubpassDependency2-srcAccessMask",
-                             "%s is non-zero when a VkMemoryBarrier2KHR is present in pNext.",
-                             loc.dot(Field::srcAccessMask).Message().c_str());
-        }
-        if (dependency.dstAccessMask != 0) {
-            skip |= LogError(objects, "UNASSIGNED-CoreChecks-VkSubpassDependency2-dstAccessMask",
-                             "%s dstAccessMask is non-zero when a VkMemoryBarrier2KHR is present in pNext.",
-                             loc.dot(Field::dstAccessMask).Message().c_str());
-        }
-        if (dependency.srcStageMask != 0) {
-            skip |= LogError(objects, "UNASSIGNED-CoreChecks-VkSubpassDependency2-srcStageMask",
-                             "%s srcStageMask is non-zero when a VkMemoryBarrier2KHR is present in pNext.",
-                             loc.dot(Field::srcStageMask).Message().c_str());
-        }
-        if (dependency.dstStageMask != 0) {
-            skip |= LogError(objects, "UNASSIGNED-CoreChecks-VkSubpassDependency2-dstStageMask",
-                             "%s dstStageMask is non-zero when a VkMemoryBarrier2KHR is present in pNext.",
-                             loc.dot(Field::dstStageMask).Message().c_str());
-        }
-        loc = in_loc.dot(Field::pNext);
-        converted_barrier = *mem_barrier;
-    } else {
-        if (mem_barrier) {
-            skip |= LogError(objects, "UNASSIGNED-CoreChecks-VkSubpassDependency2-pNext",
-                             "%s a VkMemoryBarrier2KHR is present in pNext but synchronization2 is not enabled.",
-                             loc.Message().c_str());
-        }
-        // use the subpass dependency flags, upconverted into wider synchronization2 fields.
-        converted_barrier.srcStageMask = dependency.srcStageMask;
-        converted_barrier.dstStageMask = dependency.dstStageMask;
-        converted_barrier.srcAccessMask = dependency.srcAccessMask;
-        converted_barrier.dstAccessMask = dependency.dstAccessMask;
-    }
-    auto src_queue_flags = SubpassToQueueFlags(dependency.srcSubpass);
-    skip |= ValidatePipelineStage(objects, loc.dot(Field::srcStageMask), src_queue_flags, converted_barrier.srcStageMask);
-    skip |= ValidateAccessMask(objects, loc.dot(Field::srcAccessMask), src_queue_flags, converted_barrier.srcAccessMask,
-                               converted_barrier.srcStageMask);
-
-    auto dst_queue_flags = SubpassToQueueFlags(dependency.dstSubpass);
-    skip |= ValidatePipelineStage(objects, loc.dot(Field::dstStageMask), dst_queue_flags, converted_barrier.dstStageMask);
-    skip |= ValidateAccessMask(objects, loc.dot(Field::dstAccessMask), dst_queue_flags, converted_barrier.dstAccessMask,
-                               converted_barrier.dstStageMask);
-    return skip;
 }
 
 bool CoreChecks::ValidateImageViewFormatFeatures(const IMAGE_STATE *image_state, const VkFormat view_format,
